requirements.txt
# Core API & UI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.40.0

# Vector DB & Search
chromadb>=0.5.0
rank-bm25>=0.2.2

# LLM & Embeddings
openai>=1.55.0
langchain>=0.3.0
langchain-openai>=0.2.0
sentence-transformers>=3.3.0

# Evaluation
ragas>=0.1.0

# Monitoring & Logging
loguru>=0.7.0
langsmith>=0.1.0

# PDF Processing
pymupdf>=1.25.0

# Data & Validation
pydantic>=2.9.0
python-dotenv>=1.0.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Utilities
datasets>=3.0.0  # Required by ragas
numpy>=1.26.0